+++ src/main/java/lotto/domain/LottoNumbers.java
@@ -1,9 +1,9 @@
 package lotto.domain;
 
-import static lotto.domain.ErrorMessage.INVALID_LOTTO_NUMBERS;
 import static lotto.domain.LottoNumbersCondition.AMOUNT_OF_NUMBERS;
 import static lotto.domain.LottoNumbersCondition.MAXIMUM_NUMBER;
 import static lotto.domain.LottoNumbersCondition.MINIMUM_NUMBER;
+import static lotto.domain.message.ErrorMessage.INVALID_LOTTO_NUMBERS;
 
 import java.util.List;
 import java.util.Objects;
