+++ src/main/java/lotto/domain/message/RequestMessage.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.message;
 
 public enum RequestMessage {
     PAYMENT("구입금액을 입력해 주세요."),
